requirements.txt
PyQt6
python-dateutil
python-dotenv
firebase-admin
loguru
pandas
openpyxl
qtawesome
requests
